Proposal to change object requirement for Precursor reward
Given how the packed reward now gives out the rarest (and one of the most powerful, despite only being tier 6) Precursor weapon, I felt like it might be better to have the object meant to be turned in for it be something potentially rarer and more explicitly Precursor related than the obelisk. Of course you're free to disregard this if you think that the obelisk is sufficiently rare enough.